Cleaning up and reorganize client code
